Title: 사각형 채우기 3, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240407/사각형 채우기 3/rectangle-fill-3.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 #define MAX 1000000007
-int arr[1001];
+long long arr[1001];
 
 int main() {
     int n;
@@ -13,7 +13,7 @@ int main() {
     for(int i = 3; i <= n; i++){
         arr[i] = (arr[i-1]*2 + arr[i-2]*3) % MAX;
         for(int j = i-3; j >= 0; j--){
-            arr[i] = (arr[i] + arr[j]*2) % MAX ;
+            arr[i] = (arr[i] + arr[j]*2) % MAX;
         }
     }
